refactor: auto-split test_launcher_simulation.py
